Now front-end authentication check is performed during routing

+++ server/src/models/Rating.js
@@ -1,6 +1,6 @@
 module.exports = (sequelize, DataTypes) => {
   const Rating = sequelize.define('Rating', {
-    rating: DataTypes.INTEGER
+    rating: DataTypes.FLOAT
   })
 
   Rating.associate = function (models) {
